+++ application/modules/form/controllers/IndexController.php
@@ -63,7 +63,7 @@ class Form_IndexController extends Zend_Controller_Action
      */
     public function indexAction()
     {
-        $model_sites = new Dlayer_Model_Site();
+        $model_sites = new Dlayer_Model_Admin_Site();
         $model_forms = new Dlayer_Model_Admin_Form();
 
         $forms = $model_forms->forms($this->site_id);
@@ -94,7 +94,7 @@ class Form_IndexController extends Zend_Controller_Action
      */
     private function controlBar($identity_id, $site_id)
     {
-        $model_sites = new Dlayer_Model_Site();
+        $model = new Dlayer_Model_Admin_Site();
 
         $control_bar_buttons = array(
             array(
@@ -113,7 +113,7 @@ class Form_IndexController extends Zend_Controller_Action
             array(
                 'name' => 'Your websites',
                 'class' => 'default',
-                'buttons' => $model_sites->sitesForControlBar($identity_id, $site_id)
+                'buttons' => $model->sitesForControlBar($identity_id, $site_id)
             )
         );