+++ library/Dlayer/Action/ValidateSiteId.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Validate the site id
  *
@@ -6,32 +7,31 @@
  * @copyright G3D Development Limited
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
-class Dlayer_Action_ValidateSiteId extends 
-Zend_Controller_Action_Helper_Abstract
+class Dlayer_Action_ValidateSiteId extends
+    Zend_Controller_Action_Helper_Abstract
 {
-	/**
-	 * Checks the session to validate the currently stored site id, it needs to be set, exist in the database and
-	 * belong to the currently logged in user. On error the user is redirected back to the home page
-	 *
-	 * @return void
-	 */
-	function direct() 
-	{
-		$valid = FALSE;
+    /**
+     * Checks the session to validate the currently stored site id, it needs to be set, exist in the database and
+     * belong to the currently logged in user. On error the user is redirected back to the home page
+     *
+     * @return void
+     */
+    function direct()
+    {
+        $valid = false;
 
-		$session_dlayer = new Dlayer_Session();
+        $session_dlayer = new Dlayer_Session();
 
-		if($session_dlayer->siteId() != NULL) {
-			$model_sites = new Dlayer_Model_Site();
-			if($model_sites->valid($session_dlayer->siteId(), 
-			$session_dlayer->identityId()) == TRUE) {
-				$valid = TRUE;
-			}
-		}
+        if ($session_dlayer->siteId() != null) {
+            $model = new Dlayer_Model_Admin_Site();
+            if ($model->valid($session_dlayer->siteId(), $session_dlayer->identityId()) === true) {
+                $valid = true;
+            }
+        }
 
-		if($valid == FALSE) {
-			$this->_actionController->getHelper('Redirector')->gotoUrl(
-				'/dlayer/index/home');
-		}		
-	}
+        if ($valid == false) {
+            $this->_actionController->getHelper('Redirector')->gotoUrl(
+                '/dlayer/index/home');
+        }
+    }
 }