+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v1.12: In Progress
 - The Content Manager does not display deleted content items; data remains pending feature to restore.
 - The re-order code which runs after an element is deleted in the Form Builder was sorting incorrectly. [Bugfix]
 - Added an admin controller to the Content Manager. [Refactoring]
+- Added an admin controller to the default module. [Refactoring]
 - Model cleanup [Refactoring]
 
 v1.11: 10th February 2017